requirements.txt
pennylane>=0.12.0
pyyaml
